PDF Stitcher/requirements.txt
defusedxml==0.7.1
fonttools==4.46.0
fpdf2==2.7.7
Pillow==10.1.0
